requirements.txt
flask>=2.0
requests>=2.28
python-dotenv>=1.0
neo4j>=5.0
google-search-results>=2.4
Pillow>=9.0
beautifulsoup4>=4.11